+++ packages/bpk-component-table/src/BpkTable-test.js
@@ -11,4 +11,15 @@ describe('BpkTable', () => {
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
+
+  it('should render correctly with custom class', () => {
+    const tree = renderer.create(
+      <BpkTable
+        className="my-table"
+      >
+        <tbody />
+      </BpkTable>,
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });